+++ .vscode/settings.json
@@ -1,4 +1,7 @@
 {
     "eslint.enable": true,
-    "eslint.autoFixOnSave": true
+    "eslint.autoFixOnSave": true,
+    "emmet.includeLanguages": {
+        "javascript": "javascriptreact"
+    }
 }
