+++ imports/ui/components/Users/index.js
@@ -1,0 +1,3 @@
+import Users from './Users';
+
+export default Users;
